+++ src/App.tsx
@@ -2,6 +2,7 @@ import React, { useState, useEffect, useCallback } from 'react';
 import './App.scss';
 import { peopleFromServer } from './data/people';
 import { Person } from './types/Person';
+import { Dropdown } from './components/Dropdown';
 
 type Props = {
   debounce?: number;
@@ -13,7 +14,8 @@ export const App: React.FC<Props> = ({ debounce = 300, onSelected }) => {
   const [query, setQuery] = useState<string>('');
   const [filteredPeople, setFilteredPeople] =
     useState<Person[]>(peopleFromServer);
-  const [isFocused, setIsFocused] = useState<boolean>(false);
+
+  const { name, born, died } = selectedPerson || {};
 
   const handleSelected = useCallback(
     (person: Person) => {
@@ -25,19 +27,12 @@ export const App: React.FC<Props> = ({ debounce = 300, onSelected }) => {
     [onSelected],
   );
 
-  const handleFocus = useCallback(() => setIsFocused(true), []);
-
-  const handleBlur = useCallback(() => setIsFocused(false), []);
-
   const handleQuery = useCallback(
     (event: React.ChangeEvent<HTMLInputElement>) => {
       setQuery(event.target.value);
     },
     [],
   );
-
-  const showPeople =
-    isFocused && query === '' ? peopleFromServer : filteredPeople;
 
   useEffect(() => {
     const timer = setTimeout(() => {
@@ -62,39 +57,17 @@ export const App: React.FC<Props> = ({ debounce = 300, onSelected }) => {
       <main className="section is-flex is-flex-direction-column">
         <h1 className="title" data-cy="title">
           {selectedPerson
-            ? `${selectedPerson.name} (${selectedPerson.born} - ${selectedPerson.died})`
+            ? `${name} (${born} - ${died})`
             : `No selected person`}
         </h1>
 
-        <div className="dropdown is-active">
-          <div className="dropdown-trigger">
-            <input
-              type="text"
-              placeholder="Enter a part of the name"
-              className="input"
-              data-cy="search-input"
-              value={query}
-              onChange={handleQuery}
-              onFocus={handleFocus}
-              onBlur={handleBlur}
-            />
-          </div>
+        <Dropdown
+          query={query}
+          filteredPeople={filteredPeople}
+          onQueryChange={handleQuery}
+          onPersonSelect={handleSelected}
+        />
 
-          <div className="dropdown-menu" role="menu" data-cy="suggestions-list">
-            <div className="dropdown-content">
-              {showPeople.map(person => (
-                <div
-                  className="dropdown-item"
-                  data-cy="suggestion-item"
-                  key={person.slug}
-                  onClick={() => handleSelected(person)}
-                >
-                  <p className="has-text-link">{person.name}</p>
-                </div>
-              ))}
-            </div>
-          </div>
-        </div>
         {filteredPeople.length === 0 && (
           <div
             className="
